Server Connectivity Update
This is an update which resolves many of the initial connectivity issues
that were discovered during development.
New issues have arose, as detailed in the issue log.

PLEASE NOTE:
All connectivity is currently situated on http://localhost
The sql create script, as well as the .php script being used to interact
with the app server-side are included in the update.